fix: validation and logic.

+++ main/src/main/java/ru/practicum/main/category/service/CategoryAdminServiceImpl.java
@@ -9,6 +9,8 @@ import ru.practicum.main.category.exception.CategoryNotFoundException;
 import ru.practicum.main.category.exception.CategoryUniqueNameException;
 import ru.practicum.main.category.model.Category;
 import ru.practicum.main.category.repositories.CategoryRepository;
+import ru.practicum.main.event.exception.EventConflictException;
+import ru.practicum.main.event.repositories.EventRepository;
 
 import javax.transaction.Transactional;
 
@@ -21,6 +23,7 @@ import static ru.practicum.main.category.mapper.CategoryMapperUtil.toCategoryDto
 @Transactional
 public class CategoryAdminServiceImpl implements CategoryAdminService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     public CategoryDto addCategory(NewCategoryDto request) {
@@ -40,25 +43,29 @@ public class CategoryAdminServiceImpl implements CategoryAdminService {
             log.error("Category with this id does not exist {}", catId);
             throw new CategoryNotFoundException("Category with this id does not exist");
         }
+
+        if (eventRepository.existsByCategory_Id(catId)) {
+            throw new EventConflictException("Category has associated events and cannot be deleted");
+        }
+
         log.info("Deleted category with id: {}", catId);
         categoryRepository.deleteById(catId);
     }
 
     @Override
     public CategoryDto patchCategory(Long catId, NewCategoryDto request) {
-        if (!categoryRepository.existsById(catId)) {
-            log.error("Category with this id does not exist {}", catId);
-            throw new CategoryNotFoundException("Category with this id does not exist");
+        Category oldCategory = categoryRepository.findById(catId)
+                .orElseThrow(() -> new CategoryNotFoundException("Category with this id does not exist"));
+
+        if (!oldCategory.getName().equals(request.getName()) &&
+                categoryRepository.existsByName(request.getName())) {
+            log.error("Category with this name already exists: {}", request.getName());
+            throw new CategoryUniqueNameException("Category with this name already exists: " + request.getName());
         }
-//        if (categoryRepository.existsByName(request.getName())) {
-//            log.error("Category with this name does already exist {}", request.getName());
-//            throw new CategoryUniqueNameException("Category with this name does already exist" + request.getName());
-//        }
 
-        Category category = fromDtoRequest(request);
-        category.setId(catId);
+        oldCategory.setName(request.getName());
 
-        Category save = categoryRepository.save(category);
+        Category save = categoryRepository.save(oldCategory);
         log.info("Updated category with id: {}", catId);
 
         return toCategoryDto(save);
